[pkg/ipmi] add ipmi_linux
